Add day 18

+++ src/y2022/util.kt
@@ -36,8 +36,8 @@ data class Point2(val x: Int, val y: Int) {
 data class Point2L(val x: Long, val y: Long) {
     operator fun plus(other: Point2L) = p(x + other.x, y + other.y)
     operator fun minus(other: Point2L) = p(x - other.x, y - other.y)
-    operator fun times(factor: Int) = p(x * factor, y * factor)
-    operator fun div(divisor: Int) = p(x / divisor, y / divisor)
+    operator fun times(factor: Long) = p(x * factor, y * factor)
+    operator fun div(divisor: Long) = p(x / divisor, y / divisor)
     val abs get() = p(x.absoluteValue, y.absoluteValue)
     val sum get() = x + y
     override fun toString() = "($x,$y)"
